# open-api/immich-openapi-specs.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/audit/deletes": {
      "get": {
        "operationId": "getAuditDeletes",
        "parameters": [
          {
            "name": "after",
            "required": true,
            "in": "query",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "name": "entityType",
            "required": true,
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/EntityType"
            }
          },
          {
            "name": "userId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditDeletesResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ],
        "tags": [
          "Audit"
        ]
      }
    },
    "/audit/file-report": {
      "get": {
        "operationId": "getAuditFiles",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileReportDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ],
        "tags": [
          "Audit"
        ]
      }
    },
    "/audit/file-report/checksum": {
      "post": {
        "operationId": "getFileChecksums",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FileChecksumDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FileChecksumResponseDto"
                  },
                  "type": "array"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ],
        "tags": [
          "Audit"
        ]
      }
    },
    "/audit/file-report/fix": {
      "post": {
        "operationId": "fixAuditFiles",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FileReportFixDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ],
        "tags": [
          "Audit"
        ]
      }
    },
    "/auth/admin-sign-up": {
      "post": {
        "operationId": "signUpAdmin",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignUpDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "Authentication"
        ]
      }
    },
    "/auth/change-password": {
      "post": {
        "operationId": "changePassword",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ],
        "tags": [
          "Authentication"
        ]
      }
    },
    "/auth/devices": {
      "delete": {
        "operationId": "logoutAuthDevices",
        "parameters": [],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ],
        "tags": [
          "Authentication"
        ]
      },
      "get": {
        "operationId": "getAuthDevices",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AuthDeviceResponseDto"
                  },
                  "type": "array"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ],
        "tags": [
          "Authentication"
        ]
      }
    },
    "/auth/devices/{id}": {
      "delete": {
        "operationId": "logoutAuthDevice",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ],
        "tags": [
          "Authentication"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "login",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginCredentialDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "Authentication"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "logout",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogoutResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ],
        "tags": [
          "Authentication"
        ]
      }
    },
    "/auth/validateToken": {
      "post": {
        "operationId": "validateAccessToken",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidateAccessTokenResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ],
        "tags": [
          "Authentication"
        ]
      }
    },
...
    "/oauth/authorize": {
      "post": {
        "operationId": "startOAuth",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OAuthConfigDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthAuthorizeResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "OAuth"
        ]
      }
    },
    "/oauth/callback": {
      "post": {
        "operationId": "finishOAuth",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OAuthCallbackDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "OAuth"
        ]
      }
    },
    "/oauth/link": {
      "post": {
        "operationId": "linkOAuthAccount",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OAuthCallbackDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ],
        "tags": [
          "OAuth"
        ]
      }
    },
...
    "/oauth/unlink": {
      "post": {
        "operationId": "unlinkOAuthAccount",
        "parameters": [],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ],
        "tags": [
          "OAuth"
        ]
      }
    },
...
  },
  "info": {
    "title": "Immich",
    "description": "Immich API",
    "version": "1.101.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/api"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "Bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "in": "header"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "immich_access_token"
      },
      "api_key": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    },
    "schemas": {
...
      "AuditDeletesResponseDto": {
        "properties": {
          "ids": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "needsFullSync": {
            "type": "boolean"
          }
        },
        "required": [
          "ids",
          "needsFullSync"
        ],
        "type": "object"
      },
      "AuthDeviceResponseDto": {
        "properties": {
          "createdAt": {
            "type": "string"
          },
          "current": {
            "type": "boolean"
          },
          "deviceOS": {
            "type": "string"
          },
          "deviceType": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "createdAt",
          "current",
          "deviceOS",
          "deviceType",
          "id",
          "updatedAt"
        ],
        "type": "object"
      },
...
      "ChangePasswordDto": {
        "properties": {
          "newPassword": {
            "example": "password",
            "type": "string"
          },
          "password": {
            "example": "password",
            "type": "string"
          }
        },
        "required": [
          "newPassword",
          "password"
        ],
        "type": "object"
      },
...
      "EntityType": {
        "enum": [
          "ASSET",
          "ALBUM"
        ],
        "type": "string"
      },
...
      "FileChecksumDto": {
        "properties": {
          "filenames": {
            "items": {
              "type": "string"
            },
            "type": "array"
          }
        },
        "required": [
          "filenames"
        ],
        "type": "object"
      },
      "FileChecksumResponseDto": {
        "properties": {
          "checksum": {
            "type": "string"
          },
          "filename": {
            "type": "string"
          }
        },
        "required": [
          "checksum",
          "filename"
        ],
        "type": "object"
      },
      "FileReportDto": {
        "properties": {
          "extras": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "orphans": {
            "items": {
              "$ref": "#/components/schemas/FileReportItemDto"
            },
            "type": "array"
          }
        },
        "required": [
          "extras",
          "orphans"
        ],
        "type": "object"
      },
      "FileReportFixDto": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/FileReportItemDto"
            },
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "type": "object"
      },
      "FileReportItemDto": {
        "properties": {
          "checksum": {
            "type": "string"
          },
          "entityId": {
            "format": "uuid",
            "type": "string"
          },
          "entityType": {
            "$ref": "#/components/schemas/PathEntityType"
          },
          "pathType": {
            "$ref": "#/components/schemas/PathType"
          },
          "pathValue": {
            "type": "string"
          }
        },
        "required": [
          "entityId",
          "entityType",
          "pathType",
          "pathValue"
        ],
        "type": "object"
      },
...
      "LoginCredentialDto": {
        "properties": {
          "email": {
            "example": "testuser@email.com",
            "type": "string"
          },
          "password": {
            "example": "password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "type": "object"
      },
      "LoginResponseDto": {
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "isAdmin": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          },
          "profileImagePath": {
            "type": "string"
          },
          "shouldChangePassword": {
            "type": "boolean"
          },
          "userEmail": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          }
        },
        "required": [
          "accessToken",
          "isAdmin",
          "name",
          "profileImagePath",
          "shouldChangePassword",
          "userEmail",
          "userId"
        ],
        "type": "object"
      },
      "LogoutResponseDto": {
        "properties": {
          "redirectUri": {
            "type": "string"
          },
          "successful": {
            "type": "boolean"
          }
        },
        "required": [
          "redirectUri",
          "successful"
        ],
        "type": "object"
      },
...
      "OAuthAuthorizeResponseDto": {
        "properties": {
          "url": {
            "type": "string"
          }
        },
        "required": [
          "url"
        ],
        "type": "object"
      },
      "OAuthCallbackDto": {
        "properties": {
          "url": {
            "type": "string"
          }
        },
        "required": [
          "url"
        ],
        "type": "object"
      },
      "OAuthConfigDto": {
        "properties": {
          "redirectUri": {
            "type": "string"
          }
        },
        "required": [
          "redirectUri"
        ],
        "type": "object"
      },
...
      "PathEntityType": {
        "enum": [
          "asset",
          "person",
          "user"
        ],
        "type": "string"
      },
      "PathType": {
        "enum": [
          "original",
          "preview",
          "thumbnail",
          "encoded_video",
          "sidecar",
          "face",
          "profile"
        ],
        "type": "string"
      },
...
      "SignUpDto": {
        "properties": {
          "email": {
            "example": "testuser@email.com",
            "type": "string"
          },
          "name": {
            "example": "Admin",
            "type": "string"
          },
          "password": {
            "example": "password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "name",
          "password"
        ],
        "type": "object"
      },
...
      "UserAvatarColor": {
        "enum": [
          "primary",
          "pink",
          "red",
          "yellow",
          "blue",
          "green",
          "purple",
          "orange",
          "gray",
          "amber"
        ],
        "type": "string"
      },
...
      "UserResponseDto": {
        "properties": {
          "avatarColor": {
            "$ref": "#/components/schemas/UserAvatarColor"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "deletedAt": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "isAdmin": {
            "type": "boolean"
          },
          "memoriesEnabled": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          },
          "oauthId": {
            "type": "string"
          },
          "permissions": {
            "items": {
              "enum": [
                "activity.create",
                "activity.read",
                "activity.update",
                "activity.delete",
                "album.create",
                "album.read",
                "album.update",
                "album.delete",
                "asset.create",
                "asset.read",
                "asset.update",
                "asset.delete",
                "apiKey.create",
                "apiKey.read",
                "apiKey.update",
                "apiKey.delete",
                "authDevice.create",
                "authDevice.read",
                "authDevice.update",
                "authDevice.delete",
                "face.create",
                "face.read",
                "face.update",
                "face.delete",
                "library.create",
                "library.read",
                "library.update",
                "library.delete",
                "memory.create",
                "memory.read",
                "memory.update",
                "memory.delete",
                "memory.addAsset",
                "memory.removeAsset",
                "partner.create",
                "partner.read",
                "partner.update",
                "partner.delete",
                "person.create",
                "person.read",
                "person.update",
                "person.delete",
                "report.create",
                "report.read",
                "report.update",
                "report.delete",
                "sharedLink.create",
                "sharedLink.read",
                "sharedLink.update",
                "sharedLink.delete",
                "systemConfig.read",
                "systemConfig.update",
                "systemConfig.delete",
                "stack.create",
                "stack.read",
                "stack.update",
                "stack.delete",
                "tag.create",
                "tag.read",
                "tag.update",
                "tag.delete",
                "user.create",
                "user.read",
                "user.update",
                "user.delete",
                "auth.changePassword",
                "auth.oauth",
                "album.addAsset",
                "album.removeAsset",
                "album.addUser",
                "album.removeUser",
                "asset.viewThumb",
                "asset.viewPreview",
                "asset.viewOriginal",
                "asset.upload",
                "asset.download",
                "job.read",
                "job.run",
                "map.read",
                "user.readSimple",
                "user.changePassword",
                "server.read",
                "server.setup"
              ],
              "type": "string"
            },
            "type": "array"
          },
          "profileImagePath": {
            "type": "string"
          },
          "quotaSizeInBytes": {
            "format": "int64",
            "nullable": true,
            "type": "integer"
          },
          "quotaUsageInBytes": {
            "format": "int64",
            "nullable": true,
            "type": "integer"
          },
          "shouldChangePassword": {
            "type": "boolean"
          },
          "status": {
            "$ref": "#/components/schemas/UserStatus"
          },
          "storageLabel": {
            "nullable": true,
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "avatarColor",
          "createdAt",
          "deletedAt",
          "email",
          "id",
          "isAdmin",
          "name",
          "oauthId",
          "profileImagePath",
          "quotaSizeInBytes",
          "quotaUsageInBytes",
          "shouldChangePassword",
          "status",
          "storageLabel",
          "updatedAt"
        ],
        "type": "object"
      },
      "UserStatus": {
        "enum": [
          "active",
          "removing",
          "deleted"
        ],
        "type": "string"
      },
      "ValidateAccessTokenResponseDto": {
        "properties": {
          "authStatus": {
            "type": "boolean"
          }
        },
        "required": [
          "authStatus"
        ],
        "type": "object"
      },
...
    }
  }
}